% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_gwas.R
\name{b_gwas}
\alias{b_gwas}
\title{Perform GWAS for binary covariates.}
\usage{
b_gwas(plink_path, dis_snp, bp_dis_cov)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application.}

\item{dis_snp}{Prefix for binary files for the discovery dataset.}

\item{bp_dis_cov}{File path for covariate data in the discovery dataset.}

\item{temp_dir}{Directory to save output files.}
}
\value{
None. Results are saved to files.
}
\description{
Perform GWAS for binary covariates.
}
