% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{bbp_prs}
\alias{bbp_prs}
\title{Compute Polygenic Risk Scores (PRS).}
\usage{
bbp_prs(plink_path, tar_snp)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application.}

\item{tar_snp}{Prefix for binary files for the target dataset.}

\item{temp_dir}{Directory to save output files.}
}
\value{
A list containing scaled PRS values for additive, interaction, and covariate scores.
}
\description{
Compute Polygenic Risk Scores (PRS).
}
