% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_dis_cov.R
\docType{data}
\name{bp_dis_cov}
\alias{bp_dis_cov}
\title{A phenotype data from the discovery dataset containing binary phenotype trated as exposure for the proposed causal directions.}
\format{
A data frame with 800 rows and 19 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{Binary Covariate}{A binary covariate}
\item{Confounder 1}{Confounder 1 description}
\item{Confounder 2}{Confounder 2 description}
\item{Confounder 3}{Confounder 3 description}
\item{Confounder 4}{Confounder 4 description}
\item{Confounder 5}{Confounder 5 description}
\item{Confounder 6}{Confounder 6 description}
\item{Confounder 7}{Confounder 7 description}
\item{Confounder 8}{Confounder 8 description}
\item{Confounder 9}{Confounder 9 description}
\item{Confounder 10}{Confounder 10 description}
\item{Confounder 11}{Confounder 11 description}
\item{Confounder 12}{Confounder 12 description}
\item{Confounder 13}{Confounder 13 description}
\item{Confounder 14}{Confounder 14 description}
\item{Confounder 15}{Confounder 15 description}
\item{Confounder 16}{Confounder 16 description}
}
}
\usage{
data(bp_dis_cov)
}
\description{
This dataset contains covariate  for a binary outcome information of individuals in the discovery dataset,
including various confounders.
}
\keyword{datasets}
