% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_dis_phen.R
\docType{data}
\name{bp_dis_phen}
\alias{bp_dis_phen}
\title{A phenotype data from the discovery dataset containing binary phenotype trated as an outcome for the proposed causal directions.}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{Binary outcome}{A binary outcome (1=controls, 2=cases)}
}
}
\usage{
data(bp_dis_phen)
}
\description{
This dataset contains phenotype for the outcome information for individuals in the discovery dataset,
for a binary outcome treated as a binary outcome for the proposed causal diections.
}
\keyword{datasets}
