% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_tar_phen.R
\docType{data}
\name{bp_tar_phen}
\alias{bp_tar_phen}
\title{A phenotype data from the target dataset containing binary phenotype trated as outcome for the proposed causal directions.}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{Binary outcome}{A binary outcome (0=controls, 1=cases)}
}
}
\usage{
data(bp_tar_phen)
}
\description{
This dataset contains phenotype for the outcome information of individuals in the target dataset,
}
\keyword{datasets}
