% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_dis_cov.R
\docType{data}
\name{br_dis_cov}
\alias{br_dis_cov}
\title{A phenotype data file from the discovery dataset, containing the binary exposure treated as binary outcomes for the reverse direction.}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{Binary outcome}{A binary outcome (1=controls, 2=cases)}
}
}
\usage{
data(br_dis_cov)
}
\description{
This dataset contains phenotype for the outcome information for individuals in the discovery dataset,
for a binary exposure as a binary outcome in the reverse direction.
}
\keyword{datasets}
