% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_tar_cov.R
\docType{data}
\name{br_tar_cov}
\alias{br_tar_cov}
\title{A phenotype data file from the target dataset, containing the binary exposure treated as an oucome for the reverse direction of causation}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{Binary outcome}{A binary outcome (0=controls, 1=cases)}
}
}
\usage{
data(br_tar_cov)
}
\description{
This dataset contains phenotype data for the outcome information for individuals in the target dataset,
for a binary outcome as a binary outcome in the reverse directions of causations.
}
\keyword{datasets}
