% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcim_b.R
\name{gcim_bbp}
\alias{gcim_bbp}
\title{Perform Regression Analysis for GCIM.}
\usage{
gcim_bbp(bp_tar_phen, bp_tar_cov, Add_PRS, Int_PRS, Cov_PRS, confounders)
}
\arguments{
\item{bp_tar_phen}{File path for the target phenotype data.}

\item{bp_tar_cov}{File path for the target covariate data.}

\item{Additive}{Scaled additive PRS values.}

\item{Interaction}{Scaled interaction PRS values.}

\item{Covariate}{Scaled covariate PRS values.}

\item{Confounders}{Data frame of additional confounders.}
}
\value{
Summary of the regression model.
}
\description{
Perform Regression Analysis for GCIM.
}
