% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcim_q.R
\name{gcim_q}
\alias{gcim_q}
\title{Perform Regression Analysis for GCIM.}
\usage{
gcim_q(qp_tar_phen, qp_tar_cov, Add_PRS, Int_PRS, Cov_PRS, confounders)
}
\arguments{
\item{bp_tar_phen}{File path for the target phenotype data.}

\item{bp_tar_cov}{File path for the target covariate data.}

\item{Additive}{Scaled additive PRS values.}

\item{Interaction}{Scaled interaction PRS values.}

\item{Covariate}{Scaled covariate PRS values.}

\item{Confounders}{Data frame of additional confounders.}
}
\value{
Summary of the regression model.
}
\description{
Perform Regression Analysis for GCIM.
}
