% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_dis_phen.R
\docType{data}
\name{qp_dis_phen}
\alias{qp_dis_phen}
\title{Covariate Data File for Discovery Dataset with Binary Outcomes}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{quantitative outcome}{A quantitative outcome}
#' }
}
\usage{
data(qp_dis_phen)
}
\description{
This dataset contains phenotype for the outcome information for individuals in the discovery dataset,
including various confounders for a quantitative outcome.
}
\keyword{datasets}
