% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_tar_cov.R
\docType{data}
\name{qp_tar_cov}
\alias{qp_tar_cov}
\title{Covariate Data File for Discovery Dataset with Binary phenotype}
\format{
A data frame with 800 rows and 19 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{quantitative Covariate}{A quantitative covariate}
\item{Confounder 1}{Confounder 1 description}
...
\item{Confounder 16}{Confounder 16 description}
}
}
\usage{
data(qp_tar_cov)
}
\description{
This dataset contains covariate information for individuals in the target dataset,
including various confounders for a quantitative phenotype for the outcome.
}
\keyword{datasets}
