% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_dis_cov.R
\docType{data}
\name{qr_dis_cov}
\alias{qr_dis_cov}
\title{A phenotype data file from the discovery dataset, containing the quantitative exposure treated as an outcomes for the reverse direction.}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{quantitative outcome}{A quantitative outcome}
#' }
}
\usage{
data(qr_dis_cov)
}
\description{
This dataset contains phenotype for the outcome information of individuals in the discovery dataset.
}
\keyword{datasets}
