% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_dis_phen.R
\docType{data}
\name{qr_dis_phen}
\alias{qr_dis_phen}
\title{Covariate data File for discovery dataset of the outcome as quantitative covariate phenotype for the reverse direction.}
\format{
A data frame with 800 rows and 19 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{quantitative Covariate}{A quantitative covariate}
\item{Confounder 1}{Confounder 1 description}
...
\item{Confounder 16}{Confounder 16 description}
}
}
\usage{
data(qr_dis_phen)
}
\description{
This dataset contains covariate information for individuals in the discovery dataset,
including various confounders and the outcome phenotype as a quantitative covariate.
}
\keyword{datasets}
