% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_tar_cov.R
\docType{data}
\name{bp_tar_cov}
\alias{bp_tar_cov}
\title{A phenotype data from the target dataset containing binary phenotype trated as exposure for the proposed causal directions.}
\format{
A data frame with 800 rows and 19 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{Binary Covariate}{A binary covariate}
\item{Confounder 1}{Confounder 1 description}
...
\item{Confounder 16}{Confounder 16 description}
}
}
\usage{
data(bp_tar_cov)
}
\description{
This dataset contains covariate information of individuals in the target dataset,
including various confounders.
}
\keyword{datasets}
