% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_dis_phen.R
\docType{data}
\name{qp_dis_phen}
\alias{qp_dis_phen}
\title{A phenotype data from the discovery dataset containing quantitative phenotype trated as an outcome for the proposed causal directions.}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{quantitative outcome}{A quantitative outcome}
#' }
}
\usage{
data(qp_dis_phen)
}
\description{
This dataset contains phenotype for the outcome information of individuals in the discovery dataset,
}
\keyword{datasets}
