% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_tar_cov.R
\docType{data}
\name{qr_tar_cov}
\alias{qr_tar_cov}
\title{Covariate data file for target dataset of the covariate with quantitative outcomes for the reverse direction of causation}
\format{
A data frame with 800 rows and 3 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{quantitative outcome}{A qunatitative outcome}
#' }
}
\usage{
data(qr_tar_cov)
}
\description{
This dataset contains phenotype for the outcome information for individuals in the target dataset,
including various confounders as a quantitative outcome of the covariate.
}
\keyword{datasets}
