% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_tar_phen.R
\docType{data}
\name{qr_tar_phen}
\alias{qr_tar_phen}
\title{Covariate data file for discovery dataset of the covariate with quantitative phenotype for the reverse}
\format{
A data frame with 800 rows and 19 columns:
\describe{
\item{Family ID}{Family identification number}
\item{Individual ID}{Individual identification number}
\item{quantitative Covariate}{A quantitative covariate}
\item{Confounder 1}{Confounder 1 description}
...
\item{Confounder 16}{Confounder 16 description}
}
}
\usage{
data(qr_tar_phen)
}
\description{
This dataset contains covariate information for individuals in the target dataset,
including various confounders for a quantitative phenotype of the quantiative ciovariate for the reverse direction of causation.
}
\keyword{datasets}
